# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API Documentation",
        "description": "API Documentation for Point of Sale System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/categories-products": {
            "get": {
                "tags": [
                    "CategoriesProduct"
                ],
                "summary": "Get list of category products",
                "operationId": "11a37e5d7cbcf81f9265db53f1f2b4d1",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Number of records to skip",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "CategoriesProduct"
                ],
                "summary": "Create a new category product",
                "operationId": "205020a39a0c7115c3697c13bc33cd3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slug",
                                    "name",
                                    "url"
                                ],
                                "properties": {
                                    "slug": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "discount_percentage": {
                                        "type": "number"
                                    },
                                    "rating": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/categories-products/{id}": {
            "get": {
                "tags": [
                    "CategoriesProduct"
                ],
                "summary": "Get a category product by ID",
                "operationId": "aa3557ac2c6cae44bf6afb4b513e1d18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CategoriesProduct"
                ],
                "summary": "Update a category product",
                "operationId": "2d6dc13d28611d6123e439d062264a67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "slug": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "discount_percentage": {
                                        "type": "number"
                                    },
                                    "rating": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "thumbnail": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CategoriesProduct"
                ],
                "summary": "Delete a category product",
                "operationId": "170a67a244f027d680af2092a3f7a94e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}